Split off into dropdown and dropdown-native
